requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
PyYAML==6.0
requests==2.28.2
urllib3==1.26.14
xmltodict==0.13.0
